LE_Kotlin: extracted constants and moved all relevant files into "players" package
