+++ Attt_LocalEngine_Kotlin/src/main/kotlin/logic/GameRules.kt
@@ -2,6 +2,7 @@ package logic
 
 import constants.MAX_WINNING_LINE_LENGTH
 import constants.MIN_WINNING_LINE_LENGTH
+import players.PlayerProvider
 import publicApi.AtttPlayer
 
 /**
