+++ Attt_LocalEngine_Kotlin/src/main/kotlin/players/Constants.kt
@@ -1,0 +1,9 @@
+package players
+
+internal const val PLAYER_X_NAME = "PlayerX"
+internal const val PLAYER_O_NAME = "PlayerO"
+internal const val PLAYER_NONE_NAME = "PlayerNone"
+
+internal const val ID_FOR_PLAYER_X = 0 // because by convention X is the first
+internal const val ID_FOR_PLAYER_O = 1 // by convention O goes only after X
+internal const val ID_FOR_PLAYER_NONE = -1
