+++ Attt_LocalEngine_Kotlin/src/main/kotlin/players/PlayerProvider.kt
@@ -1,26 +1,13 @@
-package logic
+package players
 
-import constants.MAX_NUMBER_OF_PLAYERS
-import constants.MIN_NUMBER_OF_PLAYERS
-import elements.Player
-import constants.SYMBOL_FOR_PLAYER_NONE
-import constants.SYMBOL_FOR_PLAYER_O
-import constants.SYMBOL_FOR_PLAYER_X
+import constants.*
 import publicApi.AtttPlayer
 import utilities.Log
-
-private const val PLAYER_X_NAME = "PlayerX"
-private const val PLAYER_O_NAME = "PlayerO"
-private const val PLAYER_NONE_NAME = "PlayerNone"
-
-private const val ID_FOR_PLAYER_X = 0 // because by convention X is the first
-private const val ID_FOR_PLAYER_O = 1 // by convention O goes only after X
-private const val ID_FOR_PLAYER_NONE = -1
 
 // for now - just the replacement of the former enums use
 internal object PlayerProvider {
 
-    val None: AtttPlayer = Player(ID_FOR_PLAYER_NONE, PLAYER_NONE_NAME, SYMBOL_FOR_PLAYER_NONE) // one for all cases
+    val None: AtttPlayer = PlayerModel(ID_FOR_PLAYER_NONE, PLAYER_NONE_NAME, SYMBOL_FOR_PLAYER_NONE) // one for all cases
 
     // this is a part of inner game logic - it should be used only internally, for now there's no need to show it to a client
     internal var activePlayer: AtttPlayer = None
@@ -28,7 +15,7 @@ internal object PlayerProvider {
 
     private var numberOfPlayersInGameSession: Int = -1 // real value cannot be less than 2 and more than 90 for now
 
-    internal var playersList: MutableList<Player> = mutableListOf()
+    internal var playersList: MutableList<PlayerModel> = mutableListOf()
 
     /**
      * resets the activePlayer and creates new instances for all players for every new GameSession instance
@@ -44,11 +31,11 @@ internal object PlayerProvider {
         Log.pl("prepareNewPlayersInstances: numberOfPlayersInGameSession = $numberOfPlayersInGameSession")
         playersList = ArrayList(numberOfPlayersInGameSession)
         if (numberOfPlayersInGameSession == MIN_NUMBER_OF_PLAYERS) { // default case for a classic Crosses & Noughts game
-            playersList.add(0, Player(ID_FOR_PLAYER_X, PLAYER_X_NAME, SYMBOL_FOR_PLAYER_X)) // usually goes first
-            playersList.add(1, Player(ID_FOR_PLAYER_O, PLAYER_O_NAME, SYMBOL_FOR_PLAYER_O)) // usually goes after X
+            playersList.add(0, PlayerModel(ID_FOR_PLAYER_X, PLAYER_X_NAME, SYMBOL_FOR_PLAYER_X)) // usually goes first
+            playersList.add(1, PlayerModel(ID_FOR_PLAYER_O, PLAYER_O_NAME, SYMBOL_FOR_PLAYER_O)) // usually goes after X
         } else { // more than 2 players
             (0 until numberOfPlayersInGameSession).forEachIndexed { index, _ ->
-                playersList.add(index, Player(index))
+                playersList.add(index, PlayerModel(index))
             }
         }
         Log.pl("prepareNewPlayersInstances: playersList = $playersList")
